Add the capability to translate the T/RPDO number to a virtual one
At the moment the PDO number is bound to the PDO index that is usually
[0..3]. But if we want to use for example TPDO48 we should have 48 TPDOs
and this consumes a lot ram. To avoid this let's introduce 2 new apis:
COTPdoTranslatePdo()
CORPdoTranslatePdo()
that allow to bind a PDO to a translated number. This way we can for
example use TPDO[0] as TPDO48.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>